refactor(settings): collapse pure toggle arms and split value_display by section
Stream B of the readability branch — zero behavior change, verified in an
isolated clone (check/clippy/fmt/settings tests green).

- settings_change: the 19 pure `draft.x = !draft.x` arms collapse into one
  or-pattern arm via draft_toggle_flag; the 6 Search* toggles (SearchConfig::
  set_enabled re-normalizes sources) and BeginnerMode (tutorial restart) keep
  explicit arms, as does every cycling/normalizing arm.
- value_display moves to settings/display.rs, split into per-section helpers
  behind an exhaustive (wildcard-free) outer routing match; arm bodies are
  byte-identical, ja/ko strings included.
- SettingsTab::sections and SettingsState::sections wildcards become explicit
  variant lists (equivalent: wildcard only covered General | Keys).

Sizes: settings_reducer.rs 1228->1183, settings.rs 1369->1091, display.rs 455.
